# gh-83076: 3.8x speed improvement in (Async)Mock instantiation