--- a/Misc/NEWS.d/next/Library/2022-12-14-17-37-01.gh-issue-83076.NaYzWT.rst
+++ b/Misc/NEWS.d/next/Library/2022-12-14-17-37-01.gh-issue-83076.NaYzWT.rst
@@ -1,0 +1,1 @@
+Instantiation of `Mock()` and `AsyncMock()` is now 3.8x faster.