This PR makes several performance optimizations to instantiation of `Mock()` and `AsyncMock()`. The changes are discussed in more detail in the linked issue #83076. The changes result in a net speedup of 3.8x for creation of both kinds of mocks.

Before this PR:

```
➜ ./python -m timeit --setup="from unittest.mock import Mock" "Mock()"
2000 loops, best of 5: 97.9 usec per loop

➜ ./python -m timeit --setup="from unittest.mock import AsyncMock" "AsyncMock()"
500 loops, best of 5: 678 usec per loop
```

After this PR:

```
➜ ./python -m timeit --setup="from unittest.mock import Mock" "Mock()"
10000 loops, best of 5: 25.6 usec per loop

➜ ./python -m timeit --setup="from unittest.mock import AsyncMock" "AsyncMock()"
2000 loops, best of 5: 181 usec per loop
```

There's also one simple code cleanup here: the `AsyncMagicMixin.__init__` method was an exact duplicate of the inherited `MagicMixin.__init__`. This is unnecessary duplication, we can just inherit it instead.


<!-- gh-issue-number: gh-83076 -->
* Issue: gh-83076
<!-- /gh-issue-number -->