@cjw296 2022-12-19
@tirkarthi - if there was a serious performance regression, then I'd suggest it's worth backporting.
That said, @carljm: for your uses, would the backport work? I haven't cranked that handle in a while...