@carljm 2022-12-19
The performance regression here occurred from 3.7 to 3.8. At least for basic instantiation of a `Mock()`, this PR eliminates that regression and brings us back to 3.7 perf level.

I don't have strong feelings about backport. For work where we are running into this regression, we use a patched Python anyway, we already carried a patch for this in 3.8 and I will backport this to whatever pre-3.12 versions we need it in, so I don't need it backported. I don't know what the usual policy in Python is for backporting perf regression fixes.